requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.17
pillow==11.0.0
numpy<2.0.0
opencv-python-headless==4.10.0.84
pdf2image==1.17.0
# Paddle OCR stack (CPU)
paddlepaddle==2.6.2
paddlex==3.0.1
